conan-public: link libjpeg to freerdp/winpr (#96)
Linking libjpeg and libpng allows use to perform image conversions via
WinPR (BMP <> JPG <> PNG)

This is only enabled for Linux and (partially) macOS

On Linux, we statically link our jpg and png dependencies. To load the
library from .NET and use this features, the consumer will probably have
to pass the `RTLD_DEEPBIND` flag to `dlopen`; otherwise for example if
the consuming application tries to load an incompatible version of
libJpeg before DevolutionsRdp is opened, it will pollute the symbols.

On macOS, there is no `RTLD_DEEPBIND` equivalent so we dynamically link
to the system libraries. This may be fragile but currently this is only
for development use on that platform.