ci: try to fix macOS build (#99)
Tapping a cask in homebrew is now failing because it wants us to provide
a GitHub identity (homebrew probably got updated in the `macos-latest`
(`macos-15`) runner.

Actually; this isn't needed anymore: GitHub reverted the installed
`cmake` version back to 3.x. We don't need to pin it.

Additionally, don't try to install `ninja`. It's already installed and
creates noise in the log files.